make container launch faster on mac
rsync from a container to a volume is _really_ slow on a mac with the
amount of files that exist in a node_modules folder. So instead I opted
for a plain yarn install inside the /app folder. This means the app
folder is filled with the packages compiled in such a way that they
work on linux. Instead of using the node_modules from the parent folder
we merely use it to warm up the cache in the container so we should be
able to do a docker-compose up from a clean folder without a network
connection.

doing a yarn install before launching the docker container will still
be faster on the mac though.